Add files via upload
This is the Python part for signal processing, now the raw signal is generated by python but it looks as same as the signal I will get from the board. When I finished the Bluetooth part I will modify only one function, now our colleague who works on the database and who workes on frontend can use this part as entry to do the following work.